Move visually similar kanji into Structure details
